Clean package names
We try to have package names which correspond to module names
