+++ beans/src/main/java/fr/pilato/elasticsearch/crawler/fs/beans/Doc.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package fr.pilato.elasticsearch.crawler.fs.meta.doc;
+package fr.pilato.elasticsearch.crawler.fs.beans;
 
 import java.util.Map;
