+++ core/src/main/java/fr/pilato/elasticsearch/crawler/fs/fileabstractor/FileAbstractor.java
@@ -19,7 +19,7 @@
 
 package fr.pilato.elasticsearch.crawler.fs.fileabstractor;
 
-import fr.pilato.elasticsearch.crawler.fs.meta.settings.FsSettings;
+import fr.pilato.elasticsearch.crawler.fs.settings.FsSettings;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
