+++ core/src/main/java/fr/pilato/elasticsearch/crawler/fs/rest/ServerStatusApi.java
@@ -22,7 +22,7 @@ package fr.pilato.elasticsearch.crawler.fs.rest;
 
 import fr.pilato.elasticsearch.crawler.fs.FsCrawlerImpl;
 import fr.pilato.elasticsearch.crawler.fs.client.ElasticsearchClient;
-import fr.pilato.elasticsearch.crawler.fs.meta.settings.FsSettings;
+import fr.pilato.elasticsearch.crawler.fs.settings.FsSettings;
 
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
