+++ core/src/main/java/fr/pilato/elasticsearch/crawler/fs/settings/FsSettingsFileHandler.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package fr.pilato.elasticsearch.crawler.fs.meta.settings;
+package fr.pilato.elasticsearch.crawler.fs.settings;
 
 import fr.pilato.elasticsearch.crawler.fs.meta.MetaFileHandler;
