+++ core/src/test/java/fr/pilato/elasticsearch/crawler/fs/test/unit/FsCrawlerUtilTest.java
@@ -19,8 +19,8 @@
 
 package fr.pilato.elasticsearch.crawler.fs.test.unit;
 
-import fr.pilato.elasticsearch.crawler.fs.meta.settings.FsSettings;
-import fr.pilato.elasticsearch.crawler.fs.meta.settings.FsSettingsFileHandler;
+import fr.pilato.elasticsearch.crawler.fs.settings.FsSettings;
+import fr.pilato.elasticsearch.crawler.fs.settings.FsSettingsFileHandler;
 import fr.pilato.elasticsearch.crawler.fs.util.FsCrawlerUtil;
 import org.junit.Test;
