+++ settings/src/main/java/fr/pilato/elasticsearch/crawler/fs/settings/Ocr.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package fr.pilato.elasticsearch.crawler.fs.meta.settings;
+package fr.pilato.elasticsearch.crawler.fs.settings;
 
 public class Ocr {
     // Language dictionary to be used.
