+++ settings/src/main/java/fr/pilato/elasticsearch/crawler/fs/settings/Rest.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package fr.pilato.elasticsearch.crawler.fs.meta.settings;
+package fr.pilato.elasticsearch.crawler.fs.settings;
 
 import java.util.Locale;
