+++ tika/src/test/java/fr/pilato/elasticsearch/crawler/fs/tika/TikaDocParserTest.java
@@ -19,9 +19,9 @@
 
 package fr.pilato.elasticsearch.crawler.fs.tika;
 
-import fr.pilato.elasticsearch.crawler.fs.meta.doc.Doc;
-import fr.pilato.elasticsearch.crawler.fs.meta.settings.Fs;
-import fr.pilato.elasticsearch.crawler.fs.meta.settings.FsSettings;
+import fr.pilato.elasticsearch.crawler.fs.beans.Doc;
+import fr.pilato.elasticsearch.crawler.fs.settings.Fs;
+import fr.pilato.elasticsearch.crawler.fs.settings.FsSettings;
 import org.apache.tika.parser.external.ExternalParser;
 import org.junit.Test;
